requirements.txt
numpy
scipy
Pillow
matplotlib
pyvista
